refactor: move server under src/http/
